refactor: simplify mship_deploy.py by extracting logic
Extract actor options and caching logic from mship_deploy.py into modelship/actor_options.py.
Extract CLI parsing arguments logic into modelship/utils/cli.py.
Rename start.py to mship_deploy.py and update associated tests and references.
