+++ tests/test_cache.py
@@ -1,14 +1,13 @@
 import os
 from unittest import mock
 
-from start import _build_cache_env_vars
-
+from modelship.actor_options import build_cache_env_vars
 from modelship.utils import cache_dir, plugins_dir
 
 
 def test_build_cache_env_vars_defaults():
     with mock.patch.dict(os.environ, {}, clear=True):
-        env_vars = _build_cache_env_vars()
+        env_vars = build_cache_env_vars()
         assert env_vars["HF_HOME"] == "/.cache/huggingface"
         assert env_vars["VLLM_CACHE_ROOT"] == "/.cache/vllm"
         assert env_vars["FLASHINFER_CACHE_DIR"] == "/.cache/flashinfer"
@@ -17,7 +16,7 @@ def test_build_cache_env_vars_defaults():
 def test_build_cache_env_vars_custom_dir():
     custom_dir = "/tmp/custom_cache"
     with mock.patch.dict(os.environ, {"MSHIP_CACHE_DIR": custom_dir}, clear=True):
-        env_vars = _build_cache_env_vars()
+        env_vars = build_cache_env_vars()
         assert env_vars["HF_HOME"] == f"{custom_dir}/huggingface"
         assert env_vars["VLLM_CACHE_ROOT"] == f"{custom_dir}/vllm"
         assert env_vars["FLASHINFER_CACHE_DIR"] == f"{custom_dir}/flashinfer"
